discrete functionalities
